api/requirements.txt
fastapi
uvicorn
pillow
numpy
torch
transformers
requests
python-multipart